pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stock-research-mcp"
version = "0.1.0"
description = "MCP server for comprehensive stock research using Alpha Vantage and Finnhub APIs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "MJ" }
]
keywords = ["mcp", "stock", "finance", "trading", "alpha-vantage", "finnhub"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "respx>=0.21.0",
]

[project.scripts]
stock-research-mcp = "stock_research.server:main"

[project.urls]
Homepage = "https://github.com/mjzou/stock-research-mcp"
Repository = "https://github.com/mjzou/stock-research-mcp"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_research"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
